changes 28-12-2020

+++ app/Models/Invoices.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use App\Models\CaseInvoices;
 
 class Invoices extends Model
 {
@@ -11,7 +12,18 @@ class Invoices extends Model
     protected $table = "invoices";
 
     static function deleteRecord($id){
-        Invoices::where("id",$id)->delete();
+    	$invoice = Invoices::with('CaseInvoice')->where("id",$id)->first();
+    	if(!empty($invoice)){
+	    	if($invoice->link_to == 'case'){
+	    		CaseInvoices::deleteRecord($invoice->CaseInvoice->id);
+	    	}
+	        Invoices::where("id",$id)->delete();
+    	}
+    }
+
+    public function CaseInvoice()
+    {
+        return $this->hasOne('App\Models\CaseInvoices','invoice_id','unique_id')->with("InvoiceItems");
     }
 
 }
